+++ app/pages/about/About.mdx
@@ -61,7 +61,7 @@ On this site, you'll find written and interactive versions of many videos, along
 
 <Hilbert
   color="oklch(55% 0.15 60)"
-  className="max-md:mask-b-from-0% max-md:mask-b-to-100% md:mask-r-from-0% md:mask-r-to-50%"
+  className="opacity-50 max-md:mask-b-from-0% max-md:mask-b-to-100% md:mask-r-from-0% md:mask-r-to-50%"
 />
 
 <div className="flex items-center gap-12 max-md:flex-col">
