+++ app/pages/lessons/2017/hilbert-curve/index.mdx
@@ -11,5 +11,5 @@ credits:
 import Hilbert from "~/components/Hilbert";
 
 <section className="aspect-video">
-  <Hilbert className="bg-black text-theme" />
+  <Hilbert className="bg-black" />
 </section>
