+++ app/util/hooks.ts
@@ -18,8 +18,8 @@ import { wrap } from "comlink";
 import { isEqual, mapValues, random } from "lodash-es";
 import { UAParser } from "ua-parser-js";
 import { celebrate } from "~/components/Celebrate";
+import { frame, sleep } from "~/util/async";
 import FuzzyWorker from "~/util/fuzzy?worker";
-import { frame, sleep } from "~/util/misc";
 import { Vector } from "~/util/vector";
 
 // check if value changed from previous render
@@ -234,7 +234,7 @@ export const useEgg = () => {
       let shape = "";
       if (today.getMonth() + 1 === 3 && today.getDate() === 14) shape = "pi";
       if (today.getMonth() + 1 === 6 && today.getDate() === 28) shape = "tau";
-      // shape = "pi";
+      shape = "pi";
       if (!shape) return;
       const width = window.innerWidth / 2;
       const height = window.innerHeight / 2;
